requirements.txt
environ==1.0
py-solc-x==2.0.2
web3==6.11.4
py-dotenv==0.1
